+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Passing custom icon to UriLabel
 
+- Pass className to popover in Tooltip
+
 ## [0.13.1] - 2020-03-04
 
 - Hotfix for modals